backend_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
numpy==1.24.3
scikit-learn==1.3.2
asyncio-mqtt==0.16.1
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
